backend/pyproject.toml
[project]
name = "surakshanet-backend"
version = "0.1.0"
description = "SurakshaNet AI - Smart Predictive Public Safety System backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.4.0",
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
ai = [
    "ultralytics>=8.2.0",
    "opencv-python-headless>=4.9.0",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "librosa>=0.10.1",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501", "B008", "B027", "B905"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
